simulator.cpp uses new DemHandler readRectangleChunk function.  Changes many value types from double to float to make this work

+++ mempa/algorithm/src/search_algorithms/SearchAlgorithm.cpp
@@ -1,6 +1,6 @@
 #include"SearchAlgorithm.h"
 
-void SearchAlgorithm::setUpAlgo(std::vector<std::vector<double> > &heightmap, std::pair<int, int> startPoint, std::pair<int, int> endPoint, double maxSlope, double pixelSize) 
+void SearchAlgorithm::setUpAlgo(std::vector<std::vector<float> > &heightmap, std::pair<int, int> startPoint, std::pair<int, int> endPoint, float maxSlope, float pixelSize) 
 {
     _heightmap = heightmap;
     _startPoint = startPoint;
