+++ mempa/algorithm/src/search_algorithms/SearchAlgorithm.h
@@ -25,10 +25,10 @@ class Node {
 
 class SearchAlgorithm {
     public:
-        void setUpAlgo(std::vector<std::vector<double> > &heightmap, std::pair<int, int> startPoint, std::pair<int, int> endPoint, double maxSlope, double pixelSize);
+        void setUpAlgo(std::vector<std::vector<float> > &heightmap, std::pair<int, int> startPoint, std::pair<int, int> endPoint, float maxSlope, float pixelSize);
         std::pair<int, int> get_step();
     protected:
-    std::vector<std::vector<double>> _heightmap;
+    std::vector<std::vector<float>> _heightmap;
     std::pair<int, int> _startPoint;
     std::pair<int, int> _endPoint;
     double _maxSlope;
